feat: 상품 요약 조회
